Added protection from accidental deletion in formulas

+++ pages/manageFormula.php
@@ -160,6 +160,12 @@ if($_POST['action'] == 'addFormula'){
 //DELETE FORMULA
 if($_GET['action'] == 'delete' && $_GET['fid']){
 	$fid = mysqli_real_escape_string($conn, $_GET['fid']);
+	
+	if(mysqli_num_rows(mysqli_query($conn, "SELECT id FROM formulasMetaData WHERE fid = '$fid' AND isProtected = '1'"))){
+		echo '<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>Error:</strong> formula '.base64_decode($fid).' is protected.</div>';
+		return;
+	}
+	
 	if(mysqli_query($conn, "DELETE FROM formulas WHERE fid = '$fid'")){
 		mysqli_query($conn, "DELETE FROM formulasMetaData WHERE fid = '$fid'");
 		echo '<div class="alert alert-success alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a>Formula '.base64_decode($fid).' deleted!</div>';
